apps/crawler_api/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.14.0
python-dotenv==1.0.0
pymongo==4.6.0
psycopg2-binary==2.9.9
google-api-python-client==2.108.0
pandas==2.1.2
tqdm==4.66.1
pytz==2023.3
lxml==4.9.3
webdriver-manager==4.0.1
fastapi==0.104.1
uvicorn==0.23.2
jinja2==3.1.2
python-multipart==0.0.6
schedule==1.2.0
croniter==1.4.1 